Add feeding API
Introduce an API to return a list of all available dishes for a specific
creature (currently just returns all, but will be tailored to each
evolution in future). Just manually "serialize" the data for now.
Decide on a better strategy later.

Introduce an API to feed a creature a given dish. In future: consider
moving the feeding logic into a Dry Operation.